+++ teslaCharger/teslaCharger/Modbus/utilities.swift
@@ -28,4 +28,32 @@ class Utilities{
         let bytes = (byteOrder == .LittleEndian) ? value : value.reversed()
         return bytes.withUnsafeBytes {$0.baseAddress!.load(as: T.self)}
     }
+    
+    
+    static func toByteArray<T>(_ value: T, byteOrder: ByteOrder) -> [UInt8] {
+        var value = value
+        let valueByteArray = withUnsafeBytes(of: &value) { Array($0)}
+        return (byteOrder == .LittleEndian) ? valueByteArray : valueByteArray.reversed()
+    }
+}
+
+class ByteArray {
+    var bytes :[UInt8] = []
+    
+    init(){
+    }
+    
+    func append<T>(_ value: T, byteCount: Int ) {
+        if byteCount == 1 {
+            //bytes
+        }
+        else if byteCount == 2{
+            
+        }
+        else{
+            print("ERROR")
+        }
+    }
+    
+    
 }
